+++ src/context/PostContext.js
@@ -1,9 +1,19 @@
 import React, { useState, useContext, useEffect } from 'react';
+import { useDispatch } from 'react-redux';
+
+// Dispatch Action
+import { createPost } from '../actions/posts';
 
 // Setup Post Context
 const PostContext = React.createContext();
 
 const PostContextProvider = ({ children }) => {
+  // Invoke Redux State
+  const dispatch = useDispatch();
+
+  // Get the current user for the post's username/author
+  const currentUser = JSON.parse(localStorage.getItem('profile'));
+
   // Final form data to be submitted to the backend
   const [postData, setPostData] = useState({});
 
@@ -42,9 +52,34 @@ const PostContextProvider = ({ children }) => {
     localStorage.setItem(stepsTempLocal, JSON.stringify(instructions));
   }, [instructions, stepsTempLocal]);
 
+  // Clear all inputs to initial state
+  const clearInputs = () => setDetails(initialDetailsData);
+
+  // Create POST to send to backend
   const handlePostPublish = (e) => {
     e.preventDefault();
-    console.log('-- postData:\n', JSON.stringify(postData, null, 2));
+
+    const newPost = setPostData({
+      ...postData,
+      ...details,
+      steps: instructions,
+      breed: selectedBreed.label,
+      username: currentUser?.result?.username,
+    });
+    // add alert to notify post is published?
+
+    // Send post to backend - CREATE POST
+    dispatch(createPost({ ...postData, newPost }));
+
+    // clear inputs back to initial values
+    clearInputs();
+    // clear temporary local storage for steps
+    localStorage.removeItem(stepsTempLocal);
+    // set breed back to null
+    setSelectedBreed(null);
+
+    console.log('Post Published!\n-- newPost:\n', newPost);
+    // redirect page to the new post
   };
 
   return (
